backend/pyproject.toml
[tool.poetry]
name = "vita-ai"
version = "1.1.0-final"
description = "Plataforma SaaS Integrada de Gestão Clínica e Financeira"
authors = ["Vita.AI Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
# Drivers de banco de dados
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"
python-multipart = "^0.0.17"

# --- AI Stack (Latest) ---
langchain = "*"
langchain-community = "*"
langchain-core = "*"
google-generativeai = "^0.8.6"
httpx = "^0.28.1"
asyncpg = "^0.31.0"
greenlet = "^3.3.0"
pandas = "^2.3.3"
aiofiles = "^25.1.0"
pdfplumber = "^0.11.8"
pypdf = "^6.5.0"
langchain-openai = "^1.1.6"
langchain-google-genai = "^4.1.2"
langchain-ollama = "^1.0.1"
faiss-cpu = "^1.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"